+++ demo/MenuViewController.swift
@@ -1,9 +1,6 @@
 import UIKit
 
 final class MenuViewController: UIViewController {
-    private static let buttonPadding: CGFloat = 14.0
-    private static let cellHeight: CGFloat = 147.0
-    private static let cellWidth: CGFloat = 180.0
     private static let infoButtonPadding: CGFloat = 20
 
     // MARK: Properties
@@ -37,7 +34,6 @@ final class MenuViewController: UIViewController {
         let distanceToObjectButton = MenuItemButton(for: .distanceToObject, action: select)
         let signsDetectionButton = MenuItemButton(for: .signsDetection, action: select)
         let objectDetectorButton = MenuItemButton(for: .objectDetection, action: select)
-        let objectMappingButton = MenuItemButton(for: .map, action: select)
         let arRoutingButton = MenuItemButton(for: .arRouting, action: select)
 
         let titleView = UIImageView(image: Asset.Assets.title.image)
@@ -54,43 +50,53 @@ final class MenuViewController: UIViewController {
         let rightLine = UIImageView(image: lineLeftRightImage)
         addToView(rightLineImageView: rightLine, withDependencyOn: centerLine)
 
-        let lineTopBottomImage = Asset.Assets.lineTopBottom.image
-        let topFirstLine = UIImageView(image: lineTopBottomImage)
-        addToView(topFirstLineImageView: topFirstLine, withDependencyOn: centerLine)
+        let topStack = createStack(subviews: [signsDetectionButton, segmentationButton, objectDetectorButton],
+                                   alignment: .bottom,
+                                   separator: verticalTopLine)
+        let bottomStack = createStack(subviews: [distanceToObjectButton, arRoutingButton, laneDetectionButton],
+                                      alignment: .top,
+                                      separator: verticalBottomLine)
 
-        let topSecondLine = UIImageView(image: lineTopBottomImage)
-        addToView(topSecondLineImageView: topSecondLine, withDependencyOn: centerLine)
-
-        let topThirdLine = UIImageView(image: lineTopBottomImage)
-        addToView(topThirdLineImageView: topThirdLine, withDependencyOn: centerLine)
-
-        addToView(signsDetectionButton: signsDetectionButton, withDependencyOn: centerLine, withDependencyOn: leftLine)
-        addToView(segmentationButton: segmentationButton, withDependencyOn: centerLine, withDependencyOn: topFirstLine)
-        addToView(objectDetectorButton: objectDetectorButton, withDependencyOn: centerLine, withDependencyOn: topSecondLine)
-        addToView(arRoutingButton: arRoutingButton, withDependencyOn: centerLine, withDependencyOn: leftLine)
-
-        let bottomFirstLine = UIImageView(image: lineTopBottomImage)
-        addToView(bottomFirstLine: bottomFirstLine, withDependencyOn: centerLine, withDependencyOn: arRoutingButton)
-
-        let bottomSecondLine = UIImageView(image: lineTopBottomImage)
-        addToView(bottomSecondLine: bottomSecondLine, withDependencyOn: centerLine, withDependencyOn: arRoutingButton)
-
-        addToView(distanceToObjectButton: distanceToObjectButton,
-                  withDependencyOn: centerLine,
-                  withDependencyOn: bottomFirstLine,
-                  withDependencyOn: leftLine)
-
-        addToView(objectMappingButton: objectMappingButton, withDependencyOn: centerLine, withDependencyOn: topThirdLine)
-
-        addToView(laneDetectionButton: laneDetectionButton,
-                  withDependencyOn: centerLine,
-                  withDependencyOn: bottomSecondLine,
-                  withDependencyOn: leftLine)
+        layout(topStack: topStack, bottomStack: bottomStack, withDependencyOn: centerLine)
 
         addToViewInfoButton()
     }
 
     // MARK: Private functions
+
+    private func createStack(
+        subviews: [UIView],
+        alignment: UIStackView.Alignment,
+        separator: () -> UIView
+    ) -> UIStackView {
+        let stack = UIStackView()
+        stack.translatesAutoresizingMaskIntoConstraints = false
+        stack.axis = .horizontal
+        stack.distribution = .equalCentering
+        stack.alignment = alignment
+
+        view.addSubview(stack)
+
+        for subview in subviews {
+            stack.addArrangedSubview(subview)
+            subview.widthAnchor.constraint(equalTo: view.safeAreaLayoutGuide.widthAnchor, multiplier: 0.25).isActive = true
+
+            if subview != subviews.last {
+                stack.addArrangedSubview(separator())
+            }
+        }
+        return stack
+    }
+
+    private func verticalTopLine() -> UIImageView {
+        return UIImageView(image: Asset.Assets.lineTopBottom.image)
+    }
+
+    private func verticalBottomLine() -> UIImageView {
+        let verticalBottomLine = UIImageView(image: Asset.Assets.lineTopBottom.image)
+        verticalBottomLine.transform = CGAffineTransform(rotationAngle: CGFloat(Double.pi))
+        return verticalBottomLine
+    }
 
     private func addToView(titleImageView: UIImageView) {
         titleImageView.translatesAutoresizingMaskIntoConstraints = false
@@ -132,158 +138,12 @@ final class MenuViewController: UIViewController {
         ])
     }
 
-    private func addToView(topFirstLineImageView: UIImageView, withDependencyOn centerLine: UIImageView) {
-        topFirstLineImageView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(topFirstLineImageView)
+    private func layout(topStack: UIStackView, bottomStack: UIStackView, withDependencyOn centerLine: UIImageView) {
         NSLayoutConstraint.activate([
-            topFirstLineImageView.bottomAnchor.constraint(equalTo: centerLine.topAnchor),
-            topFirstLineImageView.leadingAnchor.constraint(equalTo: centerLine.leadingAnchor),
-        ])
-    }
-
-    private func addToView(topSecondLineImageView: UIImageView, withDependencyOn centerLine: UIImageView) {
-        topSecondLineImageView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(topSecondLineImageView)
-        NSLayoutConstraint.activate([
-            topSecondLineImageView.bottomAnchor.constraint(equalTo: centerLine.topAnchor),
-            topSecondLineImageView.trailingAnchor.constraint(equalTo: centerLine.centerXAnchor),
-        ])
-    }
-
-    private func addToView(topThirdLineImageView: UIImageView, withDependencyOn centerLine: UIImageView) {
-        topThirdLineImageView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(topThirdLineImageView)
-        NSLayoutConstraint.activate([
-            topThirdLineImageView.bottomAnchor.constraint(equalTo: centerLine.topAnchor),
-            topThirdLineImageView.trailingAnchor.constraint(equalTo: centerLine.trailingAnchor),
-        ])
-    }
-
-    private func addToView(
-        signsDetectionButton: MenuItemButton,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn leftLine: UIImageView
-    ) {
-        view.addSubview(signsDetectionButton)
-        NSLayoutConstraint.activate([
-            signsDetectionButton.bottomAnchor.constraint(equalTo: centerLine.topAnchor),
-            signsDetectionButton.trailingAnchor.constraint(equalTo: centerLine.leadingAnchor),
-            signsDetectionButton.widthAnchor.constraint(equalTo: leftLine.widthAnchor),
-            signsDetectionButton.heightAnchor.constraint(equalToConstant: MenuViewController.cellHeight),
-        ])
-    }
-
-    private func addToView(
-        segmentationButton: MenuItemButton,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn topFirstLine: UIImageView
-    ) {
-        view.addSubview(segmentationButton)
-        NSLayoutConstraint.activate([
-            segmentationButton.bottomAnchor.constraint(equalTo: centerLine.topAnchor),
-            segmentationButton.leadingAnchor.constraint(equalTo: topFirstLine.leadingAnchor),
-            segmentationButton.widthAnchor.constraint(equalTo: centerLine.widthAnchor, multiplier: 0.5),
-            segmentationButton.heightAnchor.constraint(equalToConstant: MenuViewController.cellHeight),
-        ])
-    }
-
-    private func addToView(
-        objectDetectorButton: MenuItemButton,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn topSecondLine: UIImageView
-    ) {
-        view.addSubview(objectDetectorButton)
-        NSLayoutConstraint.activate([
-            objectDetectorButton.bottomAnchor.constraint(equalTo: centerLine.topAnchor),
-            objectDetectorButton.leadingAnchor.constraint(equalTo: topSecondLine.leadingAnchor),
-            objectDetectorButton.widthAnchor.constraint(equalTo: centerLine.widthAnchor, multiplier: 0.5),
-            objectDetectorButton.heightAnchor.constraint(equalToConstant: MenuViewController.cellHeight),
-        ])
-    }
-
-    private func addToView(
-        arRoutingButton: MenuItemButton,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn leftLine: UIImageView
-    ) {
-        view.addSubview(arRoutingButton)
-        NSLayoutConstraint.activate([
-            arRoutingButton.topAnchor.constraint(equalTo: centerLine.bottomAnchor),
-            arRoutingButton.centerXAnchor.constraint(equalTo: centerLine.centerXAnchor),
-            arRoutingButton.widthAnchor.constraint(equalTo: leftLine.widthAnchor),
-            arRoutingButton.heightAnchor.constraint(equalToConstant: MenuViewController.cellHeight),
-        ])
-    }
-
-    private func addToView(
-        bottomFirstLine: UIImageView,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn arRoutingButton: MenuItemButton
-    ) {
-        bottomFirstLine.transform = CGAffineTransform(rotationAngle: CGFloat(Double.pi))
-        bottomFirstLine.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(bottomFirstLine)
-        NSLayoutConstraint.activate([
-            bottomFirstLine.topAnchor.constraint(equalTo: centerLine.bottomAnchor),
-            bottomFirstLine.leadingAnchor.constraint(equalTo: arRoutingButton.leadingAnchor),
-        ])
-    }
-
-    private func addToView(
-        bottomSecondLine: UIImageView,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn arRoutingButton: MenuItemButton
-    ) {
-        bottomSecondLine.transform = CGAffineTransform(rotationAngle: CGFloat(Double.pi))
-        bottomSecondLine.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(bottomSecondLine)
-        NSLayoutConstraint.activate([
-            bottomSecondLine.topAnchor.constraint(equalTo: centerLine.bottomAnchor),
-            bottomSecondLine.trailingAnchor.constraint(equalTo: arRoutingButton.trailingAnchor),
-        ])
-    }
-
-    private func addToView(
-        distanceToObjectButton: MenuItemButton,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn bottomFirstLine: UIImageView,
-        withDependencyOn leftLine: UIImageView
-    ) {
-        view.addSubview(distanceToObjectButton)
-        NSLayoutConstraint.activate([
-            distanceToObjectButton.topAnchor.constraint(equalTo: centerLine.bottomAnchor),
-            distanceToObjectButton.trailingAnchor.constraint(equalTo: bottomFirstLine.leadingAnchor),
-            distanceToObjectButton.widthAnchor.constraint(equalTo: leftLine.widthAnchor),
-            distanceToObjectButton.heightAnchor.constraint(equalToConstant: MenuViewController.cellHeight),
-        ])
-    }
-
-    private func addToView(
-        objectMappingButton: MenuItemButton,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn topThirdLine: UIImageView
-    ) {
-        view.addSubview(objectMappingButton)
-        NSLayoutConstraint.activate([
-            objectMappingButton.bottomAnchor.constraint(equalTo: centerLine.topAnchor),
-            objectMappingButton.leadingAnchor.constraint(equalTo: topThirdLine.leadingAnchor),
-            objectMappingButton.widthAnchor.constraint(equalTo: centerLine.widthAnchor, multiplier: 0.5),
-            objectMappingButton.heightAnchor.constraint(equalToConstant: MenuViewController.cellHeight),
-        ])
-    }
-
-    private func addToView(
-        laneDetectionButton: MenuItemButton,
-        withDependencyOn centerLine: UIImageView,
-        withDependencyOn bottomSecondLine: UIImageView,
-        withDependencyOn leftLine: UIImageView
-    ) {
-        view.addSubview(laneDetectionButton)
-        NSLayoutConstraint.activate([
-            laneDetectionButton.topAnchor.constraint(equalTo: centerLine.bottomAnchor),
-            laneDetectionButton.leadingAnchor.constraint(equalTo: bottomSecondLine.leadingAnchor),
-            laneDetectionButton.widthAnchor.constraint(equalTo: leftLine.widthAnchor),
-            laneDetectionButton.heightAnchor.constraint(equalToConstant: MenuViewController.cellHeight),
+            topStack.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            topStack.bottomAnchor.constraint(equalTo: centerLine.topAnchor),
+            bottomStack.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            bottomStack.topAnchor.constraint(equalTo: centerLine.bottomAnchor),
         ])
     }
 
@@ -332,8 +192,6 @@ extension Screen {
             return L10n.menuSignDetectionButton
         case .objectDetection:
             return L10n.menuObjectDetectionButton
-        case .map:
-            return L10n.menuObjectMappingButton
         case .arRouting:
             return L10n.menuARButton
         case .menu:
@@ -351,8 +209,6 @@ extension Screen {
             return Asset.Assets.collisionDetection.image
         case .signsDetection:
             return Asset.Assets.icon4.image
-        case .map:
-            return Asset.Assets.icon5.image
         case .objectDetection:
             return Asset.Assets.icon6.image
         case .arRouting:
@@ -364,7 +220,8 @@ extension Screen {
 }
 
 private class MenuItemButton: UIView {
-    private let padding: CGFloat = 8
+    private static let padding: CGFloat = 8
+    private static let cellHeight: CGFloat = 147.0
 
     private let action: () -> Void
 
@@ -373,19 +230,17 @@ private class MenuItemButton: UIView {
 
         super.init(frame: .zero)
 
-        self.translatesAutoresizingMaskIntoConstraints = false
+        translatesAutoresizingMaskIntoConstraints = false
 
         addSubview(imageView)
-        NSLayoutConstraint.activate([
-            imageView.centerXAnchor.constraint(equalTo: centerXAnchor),
-            imageView.centerYAnchor.constraint(equalTo: centerYAnchor, constant: -padding),
-        ])
-
         addSubview(titleLabel)
         NSLayoutConstraint.activate([
+            imageView.centerXAnchor.constraint(equalTo: centerXAnchor),
+            imageView.centerYAnchor.constraint(equalTo: centerYAnchor, constant: -MenuItemButton.padding),
             titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
             titleLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
-            titleLabel.topAnchor.constraint(equalTo: imageView.bottomAnchor, constant: padding),
+            titleLabel.topAnchor.constraint(equalTo: imageView.bottomAnchor, constant: MenuItemButton.padding),
+            heightAnchor.constraint(equalToConstant: MenuItemButton.cellHeight),
         ])
 
         titleLabel.text = screen.title